ISSUE-10 Fix call to execFile with a file larger than 200KB will fail
Update the execFile call to specify the maxBuffer size to be tuneable
and larger than 200KB.  By default the plugin will now look up the
current file size and allocate 2x the buffer to allow room for
whitespace and formatting expansion.

This also fixes the issue where the plugin could not format files
that were large than 200KB.

Add a new configuation option to allow a user to tune the buffer
multiplier to custom values.

Update documentation.